Add a script to collect various Puppet reports
This bats script installs puppetserver and uses the same host to run
some tests using Puppet.

It uses detailed exit codes to verify the run happens correctly and uses
curl as a minimal tool to verify the result.

All reports are gathered in a created directory where each test has a
number.